# FIMEFavor.API(Controller)/FIMEFavor.API/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FIMEFavor.API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Mochila": {
      "get": {
        "tags": [
          "Mochila"
        ],
        "operationId": "GetMochila",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cuentaId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MochilaConInfoProducto"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Mochila"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cuentaId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Mochila"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Mochila/{mochilaId}": {
      "put": {
        "tags": [
          "Mochila"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "agenteId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "shoppingCartRecordId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "mochilaId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Mochila"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Mochila/{mochilaId}/{timeStamp}": {
      "delete": {
        "tags": [
          "Mochila"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "agenteId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "carritoRecordId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "timeStamp",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "mochilaId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Mochila"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Orden": {
      "get": {
        "tags": [
          "Orden"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Orden"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orden"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Orden"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Orden"
            }
          }
        }
      }
    },
    "/api/Orden/{id}": {
      "get": {
        "tags": [
          "Orden"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ordenId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Orden"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Orden"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Orden"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Productos": {
      "get": {
        "tags": [
          "Productos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Producto"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Productos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Producto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Producto"
            }
          }
        }
      }
    },
    "/api/Productos/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Productos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Cuentas": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "matricula": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "contrasena": {
          "type": "string"
        }
      }
    },
    "Producto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "modelo": {
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "categoriaId": {
          "format": "int32",
          "type": "integer"
        },
        "categoria": {
          "$ref": "#/definitions/Categoria"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "existencia": {
          "format": "int32",
          "type": "integer"
        },
        "vendedorId": {
          "format": "int32",
          "type": "integer"
        },
        "cuentas": {
          "$ref": "#/definitions/Cuentas"
        }
      }
    },
    "Categoria": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "categoriaNombre": {
          "type": "string"
        },
        "productos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Producto"
          }
        }
      }
    },
...
    "Orden": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "cuentas": {
          "$ref": "#/definitions/Cuentas"
        },
        "total": {
          "format": "double",
          "type": "number"
        },
        "lugarDeEntrega": {
          "type": "string"
        },
        "repartidorId": {
          "format": "int32",
          "type": "integer"
        },
        "cuenta": {
          "$ref": "#/definitions/Cuentas"
        },
        "detallesOrdenes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DetallesOrden"
          }
        }
      }
    },
    "DetallesOrden": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "productoId": {
          "format": "int32",
          "type": "integer"
        },
        "producto": {
          "$ref": "#/definitions/Producto"
        },
        "ordenId": {
          "format": "int32",
          "type": "integer"
        },
        "orden": {
          "$ref": "#/definitions/Orden"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        },
        "costoTotal": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "MochilaConInfoProducto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "productoId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "modelo": {
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "categoriaId": {
          "format": "int32",
          "type": "integer"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "existencia": {
          "format": "int32",
          "type": "integer"
        },
        "vendedorId": {
          "format": "int32",
          "type": "integer"
        },
        "categoriaNombre": {
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        },
        "costoTotal": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "Mochila": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "fechaCreacion": {
          "format": "date-time",
          "type": "string"
        },
        "productoId": {
          "format": "int32",
          "type": "integer"
        },
        "producto": {
          "$ref": "#/definitions/Producto"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        },
        "costoTotal": {
          "format": "double",
          "type": "number"
        },
        "cuentaId": {
          "format": "int32",
          "type": "integer"
        },
        "cuentas": {
          "$ref": "#/definitions/Cuentas"
        }
      }
    },
...
  }
}